requirements.txt
click==6.7
dnspython==1.15.0
Flask==1.0.2
flatten-json==0.1.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
Werkzeug==0.14.1
